Various bits of smarts to make administrators easier to deal with.

+++ src/java/com/stackframe/sarariman/Administrators.java
@@ -8,15 +8,17 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.AbstractCollection;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Iterator;
 import java.util.Map;
 
 /**
  *
  * @author mcculley
  */
-public class Administrators {
+public class Administrators extends AbstractCollection<Employee> {
 
     private final Sarariman sarariman;
 
@@ -24,7 +26,7 @@ public class Administrators {
         this.sarariman = sarariman;
     }
 
-    public Collection<Employee> getAdministrators() throws SQLException {
+    private Collection<Employee> getAdministrators() throws SQLException {
         Connection connection = sarariman.getConnection();
         Map<Object, Employee> employees = sarariman.getDirectory().getByNumber();
         PreparedStatement ps = connection.prepareStatement("SELECT * FROM administrators");
@@ -44,16 +46,78 @@ public class Administrators {
         }
     }
 
-    public void add(Employee employee) throws SQLException {
-        PreparedStatement ps = sarariman.getConnection().prepareStatement("INSERT INTO administrators (employee) VALUES(?)");
-        ps.setLong(1, employee.getNumber());
-        ps.executeUpdate();
+    public Iterator<Employee> iterator() {
+        try {
+            final Iterator<Employee> administrators = getAdministrators().iterator();
+            return new Iterator<Employee>() {
+
+                public boolean hasNext() {
+                    return administrators.hasNext();
+                }
+
+                public Employee next() {
+                    return administrators.next();
+                }
+
+                public void remove() {
+                    throw new UnsupportedOperationException("Not supported yet.");
+                }
+
+            };
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
     }
 
-    public void remove(Employee employee) throws SQLException {
-        PreparedStatement ps = sarariman.getConnection().prepareStatement("DELETE FROM administrators WHERE employee=?");
-        ps.setLong(1, employee.getNumber());
-        ps.executeUpdate();
+    public int size() {
+        Connection connection = sarariman.getConnection();
+        try {
+            PreparedStatement ps = connection.prepareStatement("SELECT COUNT(*) as numrows FROM administrators");
+            try {
+                ResultSet resultSet = ps.executeQuery();
+                try {
+                    resultSet.next();
+                    return resultSet.getInt("numrows");
+                } finally {
+                    resultSet.close();
+                }
+            } finally {
+                ps.close();
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public boolean add(Employee employee) {
+        try {
+            PreparedStatement ps = sarariman.getConnection().prepareStatement("INSERT INTO administrators (employee) VALUES(?)");
+            ps.setLong(1, employee.getNumber());
+            ps.executeUpdate();
+            ps.close();
+            return true;
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public boolean remove(Object o) {
+        if (!(o instanceof Employee)) {
+            return false;
+        }
+
+        Employee employee = (Employee)o;
+        try {
+            PreparedStatement ps = sarariman.getConnection().prepareStatement("DELETE FROM administrators WHERE employee=?");
+            ps.setLong(1, employee.getNumber());
+            ps.executeUpdate();
+            ps.close();
+            return true;
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
     }
 
 }
